--- src/main/java/coz/weavon/core/auth/infrastructure/adapter/AuthUserRestAdapter.java
@@ -1,57 +1,0 @@
-package coz.weavon.core.auth.infrastructure.adapter;
-
-import coz.weavon.common.exception.BusinessException;
-import coz.weavon.common.shared.Adapter;
-import coz.weavon.constant.Message;
-import coz.weavon.core.auth.application.adapter.AuthUserAdapter;
-import coz.weavon.core.auth.domain.model.AuthUser;
-import coz.weavon.core.auth.domain.model.OAuthUser;
-import coz.weavon.core.user.application.model.command.UserOperateCommand;
-import coz.weavon.core.user.application.model.command.UserSearchCommand;
-import coz.weavon.core.user.application.model.result.UserOperateResult;
-import coz.weavon.core.user.application.service.UserService;
-import coz.weavon.core.user.domain.model.User;
-import coz.weavon.core.user.domain.model.Users;
-import java.util.Optional;
-import lombok.RequiredArgsConstructor;
-
-@Adapter
-@RequiredArgsConstructor
-class AuthUserRestAdapter implements AuthUserAdapter {
-
-    private final UserService userService;
-
-    @Override
-    public Optional<AuthUser> findAuthUserByUsername(String username) {
-        UserSearchCommand searchCommand = UserSearchCommand.ofUsername(username);
-        Users foundUsers = userService.searchUsers(searchCommand);
-        return foundUsers
-                .getUserByUsername(username)
-                .map(user -> AuthUser.of(user.getUserId(), user.getUsername(), user.getPassword(), user.getRoleName()));
-    }
-
-    @Override
-    public AuthUser findAuthUserAndSaveOAuthUserIfAbsent(OAuthUser oAuthUser) {
-        UserSearchCommand searchCommand = UserSearchCommand.ofEmail(oAuthUser.getEmail());
-        Users foundUsers = userService.searchUsers(searchCommand);
-        Optional<User> optionalFoundUser = foundUsers.getUserByEmail(oAuthUser.getEmail());
-        if (optionalFoundUser.isPresent()) {
-            User foundUser = optionalFoundUser.get();
-            return AuthUser.of(
-                    foundUser.getUserId(), foundUser.getUsername(), foundUser.getPassword(), foundUser.getRoleName());
-        }
-
-        User createTargetUser = User.ofUser(oAuthUser.getEmail(), oAuthUser.getNickname(), oAuthUser.getEmail());
-        UserOperateCommand operateCommand = UserOperateCommand.ofCreateTargets(Users.of(createTargetUser));
-        UserOperateResult operateResult = userService.operateUsers(operateCommand);
-        User createdUser = operateResult
-                .getCreatedUsers()
-                .getUserByEmail(oAuthUser.getEmail())
-                .orElseThrow(() -> new BusinessException(Message.Authentication.SIGN_UP_FAILED));
-        return AuthUser.of(
-                createdUser.getUserId(),
-                createdUser.getUsername(),
-                createdUser.getPassword(),
-                createdUser.getRoleName());
-    }
-}
