+++ src/main/java/coz/weavon/core/auth/presentation/controller/AuthController.java
@@ -1,8 +1,8 @@
 package coz.weavon.core.auth.presentation.controller;
 
 import coz.weavon.common.io.RestResponse;
-import coz.weavon.core.auth.domain.model.AuthUser;
 import coz.weavon.core.auth.presentation.model.response.AuthValidResponse;
+import coz.weavon.security.model.AuthUser;
 import coz.weavon.util.AuthUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
