+++ src/main/java/coz/weavon/security/filter/UsernameAuthenticationFilter.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import coz.weavon.common.io.ErrorResponse;
 import coz.weavon.common.io.RestResponse;
 import coz.weavon.constant.Message;
-import coz.weavon.core.auth.domain.model.AuthToken;
-import coz.weavon.core.auth.domain.model.AuthUser;
 import coz.weavon.helper.MessageTranslator;
+import coz.weavon.security.model.AuthToken;
+import coz.weavon.security.model.AuthUser;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
